Quarantine unused resources
