+++ r4/src/future/main/java/gov/va/api/health/r4/api/resources/ServiceRequest.java
@@ -40,7 +40,7 @@ import lombok.NoArgsConstructor;
 )
 @Schema(
   description = "https://www.hl7.org/fhir/R4/servicerequest.html",
-  example = "SWAGGER_EXAMPLE_SERVICEREQUEST"
+  example = "serviceRequest:com.example.Example#example"
 )
 @ZeroOrOneOfs({
   @ZeroOrOneOf(
