+++ r4/src/future/main/java/gov/va/api/health/r4/api/resources/Slot.java
@@ -1,10 +1,10 @@
 package gov.va.api.health.r4.api.resources;
 
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
+import com.fasterxml.jackson.annotation.JsonProperty;
 import gov.va.api.health.r4.api.Fhir;
 import gov.va.api.health.r4.api.datatypes.CodeableConcept;
 import gov.va.api.health.r4.api.datatypes.Identifier;
-import gov.va.api.health.r4.api.datatypes.Period;
 import gov.va.api.health.r4.api.datatypes.SimpleResource;
 import gov.va.api.health.r4.api.elements.Extension;
 import gov.va.api.health.r4.api.elements.Meta;
@@ -14,7 +14,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import java.util.List;
 import javax.validation.Valid;
 import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotEmpty;
+import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Pattern;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -31,10 +31,10 @@ import lombok.NoArgsConstructor;
   isGetterVisibility = JsonAutoDetect.Visibility.NONE
 )
 @Schema(
-  description = "https://www.hl7.org/fhir/R4/schedule.html",
-  example = "SWAGGER_EXAMPLE_SCHEDULE"
+  description = "https://www.hl7.org/fhir/R4/slot.html", 
+  example = "slot:com.example.Example#example"
 )
-public class Schedule {
+public class Slot {
 
   // Ancestor -- Resource
   @Pattern(regexp = Fhir.ID)
@@ -59,22 +59,40 @@ public class Schedule {
 
   @Valid List<Extension> modifierExtension;
 
-  // Schedule Resource
+  // Slot Resource
   @Valid List<Identifier> identifier;
-
-  @Pattern(regexp = Fhir.BOOLEAN)
-  String active;
 
   @Valid List<CodeableConcept> serviceCategory;
 
   @Valid List<CodeableConcept> serviceType;
-
   @Valid List<CodeableConcept> specialty;
+  @Valid CodeableConcept appointmentType;
+  @NotNull @Valid Reference schedule;
 
-  @NotEmpty @Valid List<Reference> actor;
+  @NotNull Status status;
 
-  @Valid Period planningHorizon;
+  @NotBlank
+  @Pattern(regexp = Fhir.INSTANT)
+  String start;
+
+  @NotBlank
+  @Pattern(regexp = Fhir.INSTANT)
+  String end;
+
+  @Pattern(regexp = Fhir.BOOLEAN)
+  String overbooked;
 
   @Pattern(regexp = Fhir.STRING)
   String comment;
+
+  public enum Status {
+    busy,
+    free,
+    @JsonProperty("busy-unavailable")
+    busy_unavailable,
+    @JsonProperty("busy-tentative")
+    busy_tentative,
+    @JsonProperty("entered-in-error")
+    entered_in_error
+  }
 }
